requirements.txt
# Framework
Django>=5.0
djangorestframework
gunicorn

# Database
psycopg2-binary
SQLAlchemy

# ML/NLP
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.4.1+cu124
transformers>=4.45
datasets>=3.0.1
accelerate>=1.0.0
scikit-learn>=1.5
pandas>=2.2
pyarrow>=17
chardet>=5.2

# Data processing
pandas
numpy

# CLI/Validation/Config
typer
pydantic>=2
pyyaml